fix: resolved pr comments
