+++ packages/utils/src/index.ts
@@ -1,1 +1,2 @@
 export { default as unFormatLocaleString } from './unFormatLocaleString';
+export { default as isServiceWorkerSupported } from './isServiceWorkerSupported';
